requirements.txt
Pillow==6.1.0
pytest==5.1.2
codecov==2.0.15
pytest-cov==2.7.1
